Renamed directories and group
