+++ apis/acm/v1alpha1/register.go
@@ -16,12 +16,10 @@ limitations under the License.
 
 // NOTE: Boilerplate only. Ignore this file.
 
-// Package v1alpha1 contains API Schema definitions for the certificatemanager v1alpha1 API group
-// +k8s:openapi-gen=true
-// +k8s:deepcopy-gen=package,register
-// +k8s:conversion-gen=github.com/crossplane/provider-aws/pkg/apis/certificatemanager
-// +k8s:defaulter-gen=TypeMeta
-// +groupName=certificatemanager.aws.crossplane.io
+// Package v1alpha1 contains API Schema definitions for the acm v1alpha1 API group
+// +kubebuilder:object:generate=true
+// +groupName=acm.aws.crossplane.io
+// +versionName=v1alpha1
 package v1alpha1
 
 import (
@@ -33,7 +31,7 @@ import (
 
 // Package type metadata.
 const (
-	Group   = "certificatemanager.aws.crossplane.io"
+	Group   = "acm.aws.crossplane.io"
 	Version = "v1alpha1"
 )
