+++ apis/aws.go
@@ -20,11 +20,11 @@ package apis
 import (
 	"k8s.io/apimachinery/pkg/runtime"
 
+	acmv1alpha1 "github.com/crossplane/provider-aws/apis/acm/v1alpha1"
 	acmpcav1alpha1 "github.com/crossplane/provider-aws/apis/acmpca/v1alpha1"
 	integrationv1alpha1 "github.com/crossplane/provider-aws/apis/applicationintegration/v1alpha1"
 	cachev1alpha1 "github.com/crossplane/provider-aws/apis/cache/v1alpha1"
 	cachev1beta1 "github.com/crossplane/provider-aws/apis/cache/v1beta1"
-	certificatemanagerv1alpha1 "github.com/crossplane/provider-aws/apis/certificatemanager/v1alpha1"
 	computev1alpha3 "github.com/crossplane/provider-aws/apis/compute/v1alpha3"
 	databasev1alpha1 "github.com/crossplane/provider-aws/apis/database/v1alpha1"
 	databasev1beta1 "github.com/crossplane/provider-aws/apis/database/v1beta1"
@@ -49,7 +49,7 @@ func init() {
 		ec2v1beta1.SchemeBuilder.AddToScheme,
 		ec2v1alpha4.SchemeBuilder.AddToScheme,
 		awsv1alpha3.SchemeBuilder.AddToScheme,
-		certificatemanagerv1alpha1.SchemeBuilder.AddToScheme,
+		acmv1alpha1.SchemeBuilder.AddToScheme,
 		storagev1alpha3.SchemeBuilder.AddToScheme,
 		acmpcav1alpha1.SchemeBuilder.AddToScheme,
 		databasev1alpha1.SchemeBuilder.AddToScheme,
