+++ examples/acm/public-certificate.yaml
@@ -1,4 +1,4 @@
-apiVersion: certificatemanager.aws.crossplane.io/v1alpha1
+apiVersion: acm.aws.crossplane.io/v1alpha1
 kind: Certificate
 metadata:
   name: publiccertificate
