+++ pkg/controller/aws.go
@@ -21,12 +21,12 @@ import (
 
 	"github.com/crossplane/crossplane-runtime/pkg/logging"
 
+	"github.com/crossplane/provider-aws/pkg/controller/acm"
 	"github.com/crossplane/provider-aws/pkg/controller/acmpca/certificateauthority"
 	"github.com/crossplane/provider-aws/pkg/controller/acmpca/certificateauthoritypermission"
 	"github.com/crossplane/provider-aws/pkg/controller/applicationintegration/sqs"
 	"github.com/crossplane/provider-aws/pkg/controller/cache"
 	"github.com/crossplane/provider-aws/pkg/controller/cache/cachesubnetgroup"
-	"github.com/crossplane/provider-aws/pkg/controller/certificatemanager/certificate"
 	"github.com/crossplane/provider-aws/pkg/controller/compute"
 	"github.com/crossplane/provider-aws/pkg/controller/database"
 	"github.com/crossplane/provider-aws/pkg/controller/database/dbsubnetgroup"
@@ -87,7 +87,7 @@ func Setup(mgr ctrl.Manager, l logging.Logger) error {
 		dbsubnetgroup.SetupDBSubnetGroup,
 		certificateauthority.SetupCertificateAuthority,
 		certificateauthoritypermission.SetupCertificateAuthorityPermission,
-		certificate.SetupCertificate,
+		acm.SetupCertificate,
 		dynamodb.SetupDynamoTable,
 		sqs.SetupQueue,
 	} {
